+++ src/main/java/model/Ingredient.java
@@ -2,19 +2,19 @@ package model;
 
 public class Ingredient {
 
-    private final Integer id;
+    private final Long id;
     private final String name;
     private final Boolean inStock;
     private final Double pricePerUnit;
 
-    public Ingredient(Integer id, String name, Boolean inStock, Double pricePerUnit) {
+    public Ingredient(Long id, String name, Boolean inStock, Double pricePerUnit) {
         this.id = id;
         this.name = name;
         this.inStock = inStock;
         this.pricePerUnit = pricePerUnit;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
